added by_path and reworked into init
